Change table and model name
